@@ -13,14 +13,14 @@
   "type": "module",
   "private": true,
   "dependencies": {
-    "@angular/common": "^17.0.6",
-    "@angular/compiler": "^17.0.6",
-    "@angular/core": "^17.0.6",
-    "@angular/forms": "^17.0.6",
-    "@angular/platform-browser": "^17.0.6",
-    "@angular/platform-browser-dynamic": "^17.0.6",
-    "@angular/router": "^17.0.6",
-    "@angular/service-worker": "17.0.6",
+    "@angular/common": "^18.2.12",
+    "@angular/compiler": "^18.2.12",
+    "@angular/core": "^18.2.12",
+    "@angular/forms": "^18.2.12",
+    "@angular/platform-browser": "^18.2.12",
+    "@angular/platform-browser-dynamic": "^18.2.12",
+    "@angular/router": "^18.2.12",
+    "@angular/service-worker": "18.2.12",
     "body-parser": "^1.20.3",
     "bootstrap": "^5.3.3",
     "core-js": "^2.6.9",
@@ -29,14 +29,14 @@
     "rxjs": "6.5.4",
     "tslib": "^2.0.0",
     "web-push": "^3.3.5",
-    "zone.js": "~0.14.2"
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.0.6",
+    "@angular-devkit/build-angular": "^18.2.12",
     "@angular-eslint/builder": "^18.4.2",
     "@angular-eslint/eslint-plugin": "^18.4.2",
-    "@angular/cli": "^17.0.6",
-    "@angular/compiler-cli": "^17.0.6",
+    "@angular/cli": "^18.2.12",
+    "@angular/compiler-cli": "^18.2.12",
     "@eslint/js": "^9.15.0",
     "@types/jasmine": "~3.6.0",
     "@types/node": "^6.14.6",
@@ -52,7 +52,7 @@
     "karma-jasmine": "~4.0.0",
     "protractor": "^7.0.0",
     "ts-node": "5.0.1",
-    "typescript": "^5.2.2",
+    "typescript": "^5.5.4",
     "typescript-eslint": "^8.15.0"
   }
 }
