@@ -1,37 +1,35 @@
-import { BrowserModule } from '@angular/platform-browser';
-import { NgModule } from '@angular/core';
-import { FormsModule } from '@angular/forms';
+import {BrowserModule} from '@angular/platform-browser';
+import {NgModule} from '@angular/core';
+import {FormsModule} from '@angular/forms';
 
-import { AppComponent } from './app.component';
-import { ZipcodeEntryComponent } from './zipcode-entry/zipcode-entry.component';
-import {LocationService} from "./location.service";
-import { ForecastsListComponent } from './forecasts-list/forecasts-list.component';
-import {WeatherService} from "./weather.service";
-import { CurrentConditionsComponent } from './current-conditions/current-conditions.component';
-import { MainPageComponent } from './main-page/main-page.component';
-import {RouterModule} from "@angular/router";
-import {routing} from "./app.routing";
-import {HttpClientModule} from "@angular/common/http";
-import { ServiceWorkerModule } from '@angular/service-worker';
-import { environment } from '../environments/environment';
+import {AppComponent} from './app.component';
+import {ZipcodeEntryComponent} from './zipcode-entry/zipcode-entry.component';
+import {LocationService} from './location.service';
+import {ForecastsListComponent} from './forecasts-list/forecasts-list.component';
+import {WeatherService} from './weather.service';
+import {CurrentConditionsComponent} from './current-conditions/current-conditions.component';
+import {MainPageComponent} from './main-page/main-page.component';
+import {RouterModule} from '@angular/router';
+import {routing} from './app.routing';
+import {provideHttpClient, withInterceptorsFromDi} from '@angular/common/http';
+import {ServiceWorkerModule} from '@angular/service-worker';
+import {environment} from '../environments/environment';
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    ZipcodeEntryComponent,
-    ForecastsListComponent,
-    CurrentConditionsComponent,
-    MainPageComponent
-  ],
-  imports: [
-    BrowserModule,
-    FormsModule,
-    HttpClientModule,
-    RouterModule,
-    routing,
-    ServiceWorkerModule.register('/ngsw-worker.js', { enabled: environment.production })
-  ],
-  providers: [LocationService, WeatherService],
-  bootstrap: [AppComponent]
+    declarations: [
+        AppComponent,
+        ZipcodeEntryComponent,
+        ForecastsListComponent,
+        CurrentConditionsComponent,
+        MainPageComponent
+    ],
+    bootstrap: [AppComponent],
+    imports: [BrowserModule,
+        FormsModule,
+        RouterModule,
+        routing,
+        ServiceWorkerModule.register('/ngsw-worker.js', {enabled: environment.production})],
+    providers: [LocationService, WeatherService, provideHttpClient(withInterceptorsFromDi())]
 })
-export class AppModule { }
+export class AppModule {
+}
